+++ importer/tests/test_file_analysis.py
@@ -65,6 +65,7 @@ class MockImporter(Importer):
         return True
 
 
+@pytest.mark.skip(reason="flaky")
 @pytest.mark.skipif(
     os.environ.get("CI") == "true",
     reason="Github actions seems to not respect the memory limit",